refactor: :recycle: Pages
Refactor Page by making import more smaller though a index.ts which exports all pages components
