+++ src/pages/Error/Error.tsx
@@ -1,7 +1,7 @@
 import React from "react";
 import "./Error.scss";
 
-export default function Error() {
+export function Error() {
   return (
     <div className="Error">
       <h1>Houston, we've got a problem !</h1>
