+++ src/pages/index.ts
@@ -1,0 +1,8 @@
+export * from "./Auth/Auth";
+export * from "./Dashboard/Dashboard";
+export * from "./Error/Error";
+export * from "./Home/Home";
+export * from "./PrivateMessages/PrivateMessage";
+export * from "./Profile/Profile";
+export * from "./ServerDisplay/ServerDisplay";
+export * from "./UserDisplay/UserDisplay";
